Register freelancer backend

+++ src/main/java/DAO/EntrepriseDAO.java
@@ -11,14 +11,14 @@ public interface EntrepriseDAO {
     Boolean add(Entreprise entreprise) throws SQLException;
     
     // Read Entreprise
-    Entreprise getOneByRIB(int entreprise_RIB) throws SQLException;
+    Entreprise getOneByRIB(Long entreprise_RIB) throws SQLException;
     Set<Entreprise> getAll() throws SQLException;
     
     // Update Entreprise
     Boolean update(Entreprise entreprise) throws SQLException;
     
     // Delete Entreprise
-    Boolean deleteByRIB(int entreprise_RIB) throws SQLException;
+    Boolean deleteByRIB(Long entreprise_RIB) throws SQLException;
     
     // authentifier
     Entreprise authentifier(String email,String password) throws SQLException;
